feat: create error class with message personal

+++ src/storage/group/groupCreate.ts
@@ -1,11 +1,19 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/appError";
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroupName: string) {
   try {
-    const storageGroups = await groupsGetAll();
-    const storage = JSON.stringify([...storageGroups, newGroupName]);
+    const storadedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storadedGroups.includes(newGroupName);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo com esse nome");
+    }
+
+    const storage = JSON.stringify([...storadedGroups, newGroupName]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
